+++ src/components/ExchangeList/_components/PhotoCardInfo.jsx
